backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
alembic==1.12.1
pydantic==2.5.0
pydantic-settings==2.1.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
aiofiles==23.2.1
aiohttp==3.9.1
apscheduler==3.10.4
cryptography==41.0.8
python-dotenv==1.0.0
pandas==2.1.4
requests==2.31.0
pillow==10.1.0
psutil==5.9.6
redis==5.0.1
celery==5.3.4 